Use ORdinalEncoder correctly in split pipelines
